+++ src/io_bed.cpp
@@ -32,7 +32,7 @@ void write_bed__(const std::string& file_name,
     // Write to file
     std::string line;
     line.reserve(500);  // <- should be more than enough
-    for (uint32 i = 0; i < gamma_mats.size(); i++) {
+    for (uint64 i = 0; i < gamma_mats.size(); i++) {
         if (gamma_mats[i].n_rows == 0) continue;
         const arma::mat& gm(gamma_mats[i]);
         // First line's a bit different:
@@ -45,7 +45,7 @@ void write_bed__(const std::string& file_name,
         dbl_stream.str(std::string()); // clear stream
         out_file.write(line);
         // The rest of the lines:
-        for (uint32 j = 1; j < gm.n_rows; j++) {
+        for (uint64 j = 1; j < gm.n_rows; j++) {
             line = seq_names[i] + '\t';
             line += std::to_string(static_cast<int>(gm(j-1,0))) + '\t';
             line += std::to_string(static_cast<int>(gm(j,0))) + '\t';
